+++ public/views/draw.html
@@ -36,7 +36,7 @@
             <input type="submit" value="Finish">
         </form>
     </div>
-    <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js"></script>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
     <script type="text/javascript" src="../socket.io.min.js"></script>
     <script type="text/javascript" src="../js/draw.js"></script>
 </body>
