+++ public/views/ranking.html
@@ -24,7 +24,7 @@
         </tr>
       <% }); %>
 </table>
-    <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js"></script>
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
     <script type="text/javascript" src="../socket.io.min.js"></script>
     <script type="text/javascript" src="../js/ranking.js"></script>
 </body>
